Add unescape option (#5)
Currently, we are doing a full unescape on the XML before parsing it. But it is not a safe approach, some times you are going the find some XML's which have invalid content inside unescaped tags.

If we do the full document unescaping in this case, we are going to break the Nokogiri xpath navigation, because the invalid tags will be part of the tree, even if that tag is not being extracted.

Therefore, I removed the full unescape and added an option to just unescape the specific tags if you really need to. If there are some invalid content inside some unescaped tag content, Nokogiri can proceed with the parsing because this content will be treated as plain text content.